Added checks to see if module_name is valid in the "config chassis modules startup DPUx" cli aand also moved all the required utilities to the common file
